+++ yash-env/src/source/pretty.rs
@@ -49,21 +49,12 @@
 //! [`Report`], which then can be used in the same way to format a diagnostic
 //! message. To do this, implement `From<YourError>` or `From<&YourError>` for
 //! `Report`.
-//!
-//! ## Deprecated items
-//!
-//! Before [`Report`] was introduced, this module provided [`Message`] as the
-//! main data structure for constructing a diagnostic message. `Message` could
-//! be constructed directly or through the [`MessageBase`] trait, which
-//! provided a blanket implementation of `From<&T> for Message` for
-//! implementors of the trait. These items are now deprecated in favor of
-//! `Report`, but are available for backward compatibility for now. Please
-//! migrate to `Report` before they are removed in a future release.
 
 use super::Location;
 use std::borrow::Cow;
 use std::cell::Ref;
-use std::ops::{Deref, Range};
+use std::ops::Range;
+#[cfg(test)]
 use std::rc::Rc;
 
 /// Type of [`Report`]
@@ -235,91 +226,6 @@ impl Report<'_> {
     pub fn new() -> Self {
         Report::default()
     }
-}
-
-/// Type of annotation.
-#[deprecated(note = "Use `ReportType` or `FootnoteType` instead", since = "0.16.0")]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum AnnotationType {
-    Error,
-    Warning,
-    Info,
-    Note,
-    Help,
-}
-
-/// Source code fragment annotated with a label
-///
-/// Annotations are part of an entire [`Message`].
-#[deprecated(note = "Use `Snippet` and `Span` instead", since = "0.16.0")]
-#[derive(Clone)]
-pub struct Annotation<'a> {
-    /// Type of annotation
-    #[allow(deprecated)]
-    pub r#type: AnnotationType,
-    /// String that describes the annotated part of the source code
-    pub label: Cow<'a, str>,
-    /// Position of the annotated fragment in the source code
-    pub location: &'a Location,
-    /// Annotated code string
-    ///
-    /// This value provides an access to the string held in
-    /// `self.location.code.value`, which can only be accessed by a `Ref`.
-    pub code: Rc<dyn Deref<Target = str> + 'a>,
-}
-
-#[allow(deprecated)]
-impl std::fmt::Debug for Annotation<'_> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("Annotation")
-            .field("type", &self.r#type)
-            .field("label", &self.label)
-            .field("location", &self.location)
-            .field("code", &&**self.code)
-            .finish()
-    }
-}
-
-#[allow(deprecated)]
-impl<'a> Annotation<'a> {
-    /// Creates a new annotation.
-    ///
-    /// This function makes a borrow of `location.code.value` and stores it in
-    /// `self.code`. If it has been mutually borrowed, this function panics.
-    pub fn new(r#type: AnnotationType, label: Cow<'a, str>, location: &'a Location) -> Self {
-        Annotation {
-            r#type,
-            label,
-            location,
-            code: Rc::new(Ref::map(location.code.value.borrow(), String::as_str)),
-        }
-    }
-}
-
-/// Additional message without associated source code
-#[deprecated(note = "Use `Footnote` instead", since = "0.16.0")]
-#[derive(Clone, Debug)]
-pub struct Footer<'a> {
-    /// Type of this footer
-    #[allow(deprecated)]
-    pub r#type: AnnotationType,
-    /// Text of this footer
-    pub label: Cow<'a, str>,
-}
-
-/// Entire diagnostic message
-#[allow(deprecated)]
-#[deprecated(note = "Use `Report` instead", since = "0.16.0")]
-#[derive(Clone, Debug)]
-pub struct Message<'a> {
-    /// Type of this message
-    pub r#type: AnnotationType,
-    /// String that communicates the most important information in this message
-    pub title: Cow<'a, str>,
-    /// References to source code fragments annotated with additional information
-    pub annotations: Vec<Annotation<'a>>,
-    /// Additional text without associated source code
-    pub footers: Vec<Footer<'a>>,
 }
 
 /// Returns a mutable reference to the snippet for the given code, creating it
@@ -503,134 +409,6 @@ impl super::Source {
             }
         }
     }
-
-    /// Appends complementary annotations describing this source.
-    #[allow(deprecated)]
-    #[deprecated(note = "Use `extend_with_context` instead", since = "0.16.0")]
-    pub fn complement_annotations<'a, 's: 'a, T: Extend<Annotation<'a>>>(&'s self, result: &mut T) {
-        use super::Source::*;
-        match self {
-            Unknown
-            | Stdin
-            | CommandString
-            | CommandFile { .. }
-            | VariableValue { .. }
-            | InitFile { .. }
-            | Other { .. } => (),
-
-            CommandSubst { original } => {
-                // TODO Use Extend::extend_one
-                result.extend(std::iter::once(Annotation::new(
-                    AnnotationType::Info,
-                    "command substitution appeared here".into(),
-                    original,
-                )));
-            }
-            Arith { original } => {
-                // TODO Use Extend::extend_one
-                result.extend(std::iter::once(Annotation::new(
-                    AnnotationType::Info,
-                    "arithmetic expansion appeared here".into(),
-                    original,
-                )));
-            }
-            Eval { original } => {
-                // TODO Use Extend::extend_one
-                result.extend(std::iter::once(Annotation::new(
-                    AnnotationType::Info,
-                    "command passed to the eval built-in here".into(),
-                    original,
-                )));
-            }
-            DotScript { name, origin } => {
-                // TODO Use Extend::extend_one
-                result.extend(std::iter::once(Annotation::new(
-                    AnnotationType::Info,
-                    format!("script `{name}` was sourced here",).into(),
-                    origin,
-                )));
-            }
-            Trap { origin, .. } => {
-                // TODO Use Extend::extend_one
-                result.extend(std::iter::once(Annotation::new(
-                    AnnotationType::Info,
-                    "trap was set here".into(),
-                    origin,
-                )));
-            }
-            Alias { original, alias } => {
-                // TODO Use Extend::extend_one
-                result.extend(std::iter::once(Annotation::new(
-                    AnnotationType::Info,
-                    format!("alias `{}` was substituted here", alias.name).into(),
-                    original,
-                )));
-                original.code.source.complement_annotations(result);
-                result.extend(std::iter::once(Annotation::new(
-                    AnnotationType::Info,
-                    format!("alias `{}` was defined here", alias.name).into(),
-                    &alias.origin,
-                )));
-                alias.origin.code.source.complement_annotations(result);
-            }
-        }
-    }
-}
-
-/// Helper for constructing a [`Message`]
-///
-/// Thanks to the blanket implementation `impl<'a, T: MessageBase> From<&'a T>
-/// for Message<'a>`, implementors of this trait can be converted to a message
-/// for free.
-#[allow(deprecated)]
-#[deprecated(note = "Use `Report` instead", since = "0.16.0")]
-pub trait MessageBase {
-    /// Returns the type of the entire message.
-    ///
-    /// The default implementation returns `AnnotationType::Error`.
-    fn message_type(&self) -> AnnotationType {
-        AnnotationType::Error
-    }
-
-    // TODO message tag
-
-    /// Returns the main caption of the message.
-    fn message_title(&self) -> Cow<'_, str>;
-
-    /// Returns an annotation to be the first in the message.
-    fn main_annotation(&self) -> Annotation<'_>;
-
-    /// Adds additional annotations to the given container.
-    ///
-    /// The default implementation does nothing.
-    fn additional_annotations<'a, T: Extend<Annotation<'a>>>(&'a self, results: &mut T) {
-        let _ = results;
-    }
-
-    /// Returns footers that are included in the message.
-    fn footers(&self) -> Vec<Footer<'_>> {
-        Vec::new()
-    }
-}
-
-/// Constructs a message based on the message base.
-#[allow(deprecated)]
-impl<'a, T: MessageBase> From<&'a T> for Message<'a> {
-    fn from(base: &'a T) -> Self {
-        let main_annotation = base.main_annotation();
-        let main_source = &main_annotation.location.code.source;
-        let mut annotations = vec![main_annotation];
-
-        main_source.complement_annotations(&mut annotations);
-        base.additional_annotations(&mut annotations);
-
-        Message {
-            r#type: base.message_type(),
-            title: base.message_title(),
-            annotations,
-            footers: base.footers(),
-        }
-    }
 }
 
 mod annotate_snippets_support {
@@ -647,7 +425,7 @@ mod annotate_snippets_support {
         }
     }
 
-    /// Converts `yash_syntax::source::pretty::Span` into
+    /// Converts `yash_env::source::pretty::Span` into
     /// `annotate_snippets::Annotation`.
     ///
     /// This conversion is not provided as a public `From<&Span> for Annotation` implementation
@@ -667,7 +445,7 @@ mod annotate_snippets_support {
     // because a future variant of `SpanRole` may map to
     // `annotate_snippets::Patch` instead of `annotate_snippets::Annotation`.
 
-    /// Converts `yash_syntax::source::pretty::Snippet` into
+    /// Converts `yash_env::source::pretty::Snippet` into
     /// `annotate_snippets::Snippet<'a, annotate_snippets::Annotation<'a>>`.
     ///
     /// This conversion is not provided as a public `From<&Snippet> for Snippet` implementation
@@ -690,11 +468,8 @@ mod annotate_snippets_support {
             .annotations(snippet.spans.iter().map(span_to_annotation))
     }
 
-    /// Converts `yash_syntax::source::pretty::FootnoteType` into
+    /// Converts `yash_env::source::pretty::FootnoteType` into
     /// `annotate_snippets::Level`.
-    ///
-    /// This implementation is only available when the `yash_syntax` crate is
-    /// built with the `annotate-snippets` feature enabled.
     impl From<FootnoteType> for annotate_snippets::Level<'_> {
         fn from(r#type: FootnoteType) -> Self {
             use FootnoteType::*;
@@ -706,33 +481,24 @@ mod annotate_snippets_support {
         }
     }
 
-    /// Converts `yash_syntax::source::pretty::Footnote` into
+    /// Converts `yash_env::source::pretty::Footnote` into
     /// `annotate_snippets::Message`.
-    ///
-    /// This implementation is only available when the `yash_syntax` crate is
-    /// built with the `annotate-snippets` feature enabled.
     impl<'a> From<Footnote<'a>> for annotate_snippets::Message<'a> {
         fn from(footer: Footnote<'a>) -> Self {
             annotate_snippets::Level::from(footer.r#type).message(footer.label)
         }
     }
 
-    /// Converts `&yash_syntax::source::pretty::Footnote` into
+    /// Converts `&yash_env::source::pretty::Footnote` into
     /// `annotate_snippets::Message`.
-    ///
-    /// This implementation is only available when the `yash_syntax` crate is
-    /// built with the `annotate-snippets` feature enabled.
     impl<'a> From<&'a Footnote<'a>> for annotate_snippets::Message<'a> {
         fn from(footer: &'a Footnote<'a>) -> Self {
             annotate_snippets::Level::from(footer.r#type).message(&*footer.label)
         }
     }
 
-    /// Converts `yash_syntax::source::pretty::Report` into
+    /// Converts `yash_env::source::pretty::Report` into
     /// `annotate_snippets::Group`.
-    ///
-    /// This implementation is only available when the `yash_syntax` crate is
-    /// built with the `annotate-snippets` feature enabled.
     impl<'a> From<&'a Report<'a>> for annotate_snippets::Group<'a> {
         fn from(report: &'a Report<'a>) -> Self {
             let title = annotate_snippets::Level::from(report.r#type).primary_title(&*report.title);
@@ -750,94 +516,6 @@ mod annotate_snippets_support {
                         .iter()
                         .map(annotate_snippets::Message::from),
                 )
-        }
-    }
-
-    /// Converts `yash_syntax::source::pretty::AnnotationType` into
-    /// `annotate_snippets::Level`.
-    ///
-    /// This implementation is only available when the `yash_syntax` crate is
-    /// built with the `annotate-snippets` feature enabled.
-    #[allow(deprecated)]
-    impl<'a> From<AnnotationType> for annotate_snippets::Level<'a> {
-        fn from(r#type: AnnotationType) -> Self {
-            use AnnotationType::*;
-            match r#type {
-                Error => Self::ERROR,
-                Warning => Self::WARNING,
-                Info => Self::INFO,
-                Note => Self::NOTE,
-                Help => Self::HELP,
-            }
-        }
-    }
-
-    /// Converts `yash_syntax::source::pretty::AnnotationType` into
-    /// `annotate_snippets::AnnotationKind`.
-    ///
-    /// This implementation is only available when the `yash_syntax` crate is
-    /// built with the `annotate-snippets` feature enabled.
-    #[allow(deprecated)]
-    impl From<AnnotationType> for annotate_snippets::AnnotationKind {
-        fn from(r#type: AnnotationType) -> Self {
-            use AnnotationType::*;
-            match r#type {
-                Error | Warning => Self::Primary,
-                Info | Note | Help => Self::Context,
-            }
-        }
-    }
-
-    /// Converts `yash_syntax::source::pretty::Message` into
-    /// `annotate_snippets::Group`.
-    ///
-    /// This implementation is only available when the `yash_syntax` crate is
-    /// built with the `annotate-snippets` feature enabled.
-    #[allow(deprecated)]
-    impl<'a> From<&'a Message<'a>> for annotate_snippets::Group<'a> {
-        fn from(message: &'a Message<'a>) -> Self {
-            let mut snippets: Vec<(
-                &super::super::Code,
-                annotate_snippets::Snippet<'a, annotate_snippets::Annotation<'a>>,
-                Vec<annotate_snippets::Annotation>,
-            )> = Vec::new();
-            // We basically convert each annotation into a snippet, but want to merge annotations
-            // with the same code into a single snippet. For this, we first collect all annotations
-            // into a temporary vector, and then merge annotations with the same code into a single
-            // snippet.
-            for annotation in &message.annotations {
-                let range = annotation.location.byte_range();
-                let as_annotation = annotate_snippets::AnnotationKind::from(annotation.r#type)
-                    .span(range)
-                    .label(&annotation.label);
-                let code = &*annotation.location.code;
-                if let Some((_, _, annotations)) =
-                    snippets.iter_mut().find(|&&mut (c, _, _)| c == code)
-                {
-                    annotations.push(as_annotation);
-                } else {
-                    let line_start = code
-                        .start_line_number
-                        .get()
-                        .try_into()
-                        .unwrap_or(usize::MAX);
-                    let snippet = annotate_snippets::Snippet::source(&**annotation.code)
-                        .line_start(line_start)
-                        .path(code.source.label());
-                    snippets.push((code, snippet, vec![as_annotation]));
-                }
-            }
-
-            annotate_snippets::Level::from(message.r#type)
-                .primary_title(&*message.title)
-                .elements(
-                    snippets
-                        .into_iter()
-                        .map(|(_, snippet, annotations)| snippet.annotations(annotations)),
-                )
-                .elements(message.footers.iter().map(|footer| {
-                    annotate_snippets::Level::from(footer.r#type).message(&*footer.label)
-                }))
         }
     }
 }